+++ examples/TPMS/surface/fischerKochS.py
@@ -1,11 +1,12 @@
-from microgen import Tpms, tpms
+from microgen import Tpms
+from microgen.shape.surface_functions import fischerKochS
 import pyvista as pv
 
 geometry = Tpms(
-    surface_function=tpms.fischerKochS,
+    surface_function=fischerKochS,
     repeat_cell=5
 )
-mesh = geometry.generateSurfaceVtk()
+mesh = geometry.generateVtk(type_part="surface")
 
 pl = pv.Plotter()
 pl.add_mesh(mesh)
