+++ examples/TPMS/tpms/tpms.py
@@ -1,39 +1,45 @@
-import cadquery as cq
-from microgen import Tpms, tpms
 import numpy as np
+import pyvista as pv
+
+from microgen import Tpms
+from microgen.shape import surface_functions
 
 surfaces = [
-    tpms.neovius,
-    tpms.gyroid,
-    tpms.schwarzD,
-    tpms.schwarzP,
-    tpms.honeycomb,
-    tpms.schoenIWP,
-    tpms.schoenFRD,
-    tpms.fischerKochS,
-    tpms.pmy,
+    surface_functions.neovius,
+    surface_functions.gyroid,
+    surface_functions.schwarzD,
+    surface_functions.schwarzP,
+    surface_functions.honeycomb,
+    surface_functions.schoenIWP,
+    surface_functions.schoenFRD,
+    surface_functions.fischerKochS,
+    surface_functions.pmy,
 ]
 
-assembly = cq.Assembly()
+meshes = []
 
 i = 0
 n_col = 3
 n_row = np.ceil(len(surfaces) / n_col)
-for surface in surfaces:
+for i, surface in enumerate(surfaces):
     i_x = i % n_col
     i_y = i // n_col
 
     elem = Tpms(
-        center=(0.5, 0.5, 0.5),
-        orientation=(90, 90, 90),
         surface_function=surface,
-        type_part="sheet",
-        thickness=0.0
+        offset=0.3,
+        resolution=100,
     )
-    center = (1.2 * (i_x - 0.5 * (n_col - 1)), -1.2 * (i_y - 0.5 * (n_row - 1)), 0)
-    assembly.add(elem.generateSurface(), loc=cq.Location(cq.Vector(center)))
+    
+    # center = (1.2 * (i_x - 0.5 * (n_col - 1)), -1.2 * (i_y - 0.5 * (n_row - 1)), 0)
+    mesh = elem.sheet
+    mesh.translate(
+        [
+            1.2 * (i_x - 0.5 * (n_col - 1)),
+            -1.2 * (i_y - 0.5 * (n_row - 1)),
+            0
+        ], inplace=True
+    )
+    meshes.append(mesh)
 
-    i = i + 1
-    print(" {}/{}".format(i, len(surfaces)), end="\r")
-
-cq.exporters.export(assembly.toCompound(), "tpms.stl")
+pv.MultiBlock(meshes).plot(color="w")
